requirements.txt
absl-py==2.1.0
asttokens==3.0.0
astunparse==1.6.3
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
efficientnet_pytorch==0.7.1
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.12.23
fonttools==4.55.3
fsspec==2024.12.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.12.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
imutils==0.5.4
intel-openmp==2021.4.0
ipython==8.31.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jax==0.5.0
jaxlib==0.5.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.20
mkl==2021.4.0
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
onnx==1.17.0
onnxruntime-gpu==1.20.1
onnxslim==0.1.47
open3d==0.19.0
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pika==1.3.2
pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.49
protobuf==4.25.6
psutil==6.1.1
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
# pywin32==308
PyYAML==6.0.2
referencing==0.36.1
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rpds-py==0.22.3
scikit-learn==1.6.1
scipy==1.15.0
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==75.1.0
six==1.17.0
sounddevice==0.5.1
stack-data==0.6.3
sympy==1.13.1
tbb==2021.11.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
# tensorflow_intel==2.18.0
tensorflow==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
# torch==2.3.0+cu118
# torchaudio==2.3.0+cu118
# torchvision==0.18.0+cu118
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.58
ultralytics-thop==2.0.13
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.0.6
wheel==0.44.0
widgetsnbextension==4.0.13
wrapt==1.17.2
zipp==3.21.0
